Add explicit casts for uintptr_t printf() where needed
Since uintptr_t may be %lu (typical in U-Boot) or %u (stdint.h on machine
architectures) we cannot be sure which printf() format string to use. Add
casts where necessary to avoid compiler warnings.

+++ common/cmd_scsi.c
@@ -407,7 +407,7 @@ static ulong scsi_read(int device, ulong blknr, lbaint_t blkcnt, void *buffer)
 		}
 		debug("scsi_read_ext: startblk " LBAF
 		      ", blccnt %x buffer %lx\n",
-		      start, smallblks, buf_addr);
+		      start, smallblks, (ulong)buf_addr);
 		if (scsi_exec(pccb) != true) {
 			scsi_print_error(pccb);
 			blkcnt-=blks;
@@ -416,7 +416,7 @@ static ulong scsi_read(int device, ulong blknr, lbaint_t blkcnt, void *buffer)
 		buf_addr+=pccb->datalen;
 	} while(blks!=0);
 	debug("scsi_read_ext: end startblk " LBAF
-	      ", blccnt %x buffer %lx\n", start, smallblks, buf_addr);
+	      ", blccnt %x buffer %lx\n", start, smallblks, (ulong)buf_addr);
 	return(blkcnt);
 }
 
@@ -461,7 +461,7 @@ static ulong scsi_write(int device, ulong blknr,
 			blks = 0;
 		}
 		debug("%s: startblk " LBAF ", blccnt %x buffer %lx\n",
-		      __func__, start, smallblks, buf_addr);
+		      __func__, start, smallblks, (ulong)buf_addr);
 		if (scsi_exec(pccb) != true) {
 			scsi_print_error(pccb);
 			blkcnt -= blks;
@@ -470,7 +470,7 @@ static ulong scsi_write(int device, ulong blknr,
 		buf_addr += pccb->datalen;
 	} while (blks != 0);
 	debug("%s: end startblk " LBAF ", blccnt %x buffer %lx\n",
-	      __func__, start, smallblks, buf_addr);
+	      __func__, start, smallblks, (ulong)buf_addr);
 	return blkcnt;
 }
